+++ tests/apps/cgstats/conftest.py
@@ -5,16 +5,18 @@ import pytest
 from cg.apps.cgstats.crud import create
 from cg.apps.cgstats.stats import StatsAPI
 from cg.models.demultiplex.demux_results import DemuxResults
-from tests.apps.demultiplex.conftest import fixture_demultiplex_fixtures
-from tests.models.demultiplexing.conftest import (
-    fixture_bcl2fastq_demux_results,
-    fixture_demultiplexed_flowcell,
-    fixture_demultiplexed_runs,
-    fixture_flowcell_full_name,
-    fixture_flowcell_object,
-    fixture_flowcell_path,
-    fixture_flowcell_runs,
-)
+from tests.apps.demultiplex.conftest import *
+from tests.models.demultiplexing.conftest import *
+
+# from tests.models.demultiplexing.conftest import (
+#     fixture_bcl2fastq_demux_results,
+#     fixture_demultiplexed_flowcell,
+#     fixture_demultiplexed_runs,
+#     fixture_flowcell_full_name,
+#     fixture_flowcell_object,
+#     fixture_flowcell_path,
+#     fixture_flowcell_runs,
+# )
 
 
 @pytest.fixture(name="stats_api")
