+++ tests/models/demultiplexing/conftest.py
@@ -6,7 +6,7 @@ import pytest
 from cg.models.demultiplex.demux_results import DemuxResults
 from cg.models.demultiplex.flowcell import Flowcell
 from tests.apps.demultiplex.conftest import (
-    fixture_bcl2fastq_demux_outdir,
+    fixture_bcl2fastq_finished_flowcell,
     fixture_bcl2fastq_flowcell_dir,
     fixture_demultiplex_fixtures,
 )
@@ -42,22 +42,29 @@ def fixture_demultiplexed_flowcell(demultiplexed_runs: Path, flowcell_full_name:
     return demultiplexed_runs / flowcell_full_name
 
 
-# @pytest.fixture(name="finished_bcl2fastq_demux_dir")
-# def fixture_finish_bcl2fastq_demux_dir(finished_demuxes_dir: Path) -> Path:
-#     """returns the path to all finished bcl2fastq demux fixtures"""
-#     return finished_demuxes_dir / "bcl2fastq"
+# @pytest.fixture(name="bcl2fastq_flowcell_object")
+# def fixture_bcl2fastq_flowcell_object(bcl2fastq_flowcell_dir: Path) -> Flowcell:
+#     return Flowcell(bcl2fastq_flowcell_dir)
+#
+#
+# @pytest.fixture(name="bcl2fastq_demux_results")
+# def fixture_bcl2fastq_demux_results(
+#     demultiplexed_flowcell: Path,
+#     bcl2fastq_flowcell_object: Flowcell,
+#     bcl2fastq_demux_stats_files: Dict,
+# ) -> DemuxResults:
+#     return DemuxResults(
+#         demux_dir=demultiplexed_flowcell,
+#         flowcell=bcl2fastq_flowcell_object,
+#         bcl_converter="bcl2fastq",
+#         demux_stats_files=bcl2fastq_demux_stats_files,
+#     )
 
 
 @pytest.fixture(name="finished_dragen_demux_dir")
-def fixture_finish_bcl2fastq_demux_dir(finished_demuxes_dir: Path) -> Path:
+def fixture_finish_dragen_demux_dir(finished_demuxes_dir: Path) -> Path:
     """returns the path to all finished bcl2fastq demux fixtures"""
     return finished_demuxes_dir / "dragen"
-
-
-# @pytest.fixture(name="bcl2fastq_demux_outdir")
-# def fixture_bcl2fastq_demux_outdir(finished_bcl2fastq_demux_dir: Path) -> Path:
-#     """return the path to a finished demultiplexed run (bcl2fastq)"""
-#     return finished_bcl2fastq_demux_dir / "210428_A00689_0260_BHNC2FDSXY"
 
 
 @pytest.fixture(name="dragen_demux_outdir")
@@ -66,40 +73,16 @@ def fixture_dragen_demux_outdir(finished_dragen_demux_dir: Path) -> Path:
     return finished_dragen_demux_dir / "210428_A00689_0260_BHNC2FDSXY"
 
 
-# @pytest.fixture(name="bcl2fastq_unaligned_dir")
-# def fixture_bcl2fastq_unaligned_dir(finished_bcl2fastq_demux_dir: Path) -> Path:
-#     """return the path to a bcl2fastq unaligned dir"""
-#     return finished_bcl2fastq_demux_dir / "Unaligned"
-
-
 @pytest.fixture(name="dragen_unaligned_dir")
 def fixture_dragen_unaligned_dir(finished_dragen_demux_dir: Path) -> Path:
     """return the path to a bcl2fastq unaligned dir"""
     return finished_dragen_demux_dir / "Unaligned"
 
 
-# @pytest.fixture(name="bcl2fastq_stats_dir")
-# def fixture_bcl2fastq_stats_dir(bcl2fastq_unaligned_dir: Path) -> Path:
-#     """return the path to a bcl2fastq Stats dir"""
-#     return bcl2fastq_unaligned_dir / "Stats"
-
-
 @pytest.fixture(name="dragen_reports_dir")
 def fixture_dragen_reports_dir(dragen_unaligned_dir: Path) -> Path:
     """return the path to a Dragen Reports dir"""
     return dragen_unaligned_dir / "Reports"
-
-
-# @pytest.fixture(name="bcl2fastq_conversion_stats")
-# def fixture_bcl2fastq_conversion_stats(bcl2fastq_stats_dir: Path) -> Path:
-#     """return the path to the conversion stats file"""
-#     return bcl2fastq_stats_dir / "ConversionStats.xml"
-#
-#
-# @pytest.fixture(name="bcl2fastq_demultiplexing_stats")
-# def fixture_bcl2fastq_demultiplexing_stats(bcl2fastq_stats_dir: Path) -> Path:
-#     """return the path to the demultiplexing stats file"""
-#     return bcl2fastq_stats_dir / "DemultiplexingStats.xml"
 
 
 @pytest.fixture(name="dragen_adapter_metrics")
@@ -120,18 +103,6 @@ def fixture_dragen_run_info(dragen_reports_dir: Path) -> Path:
     return dragen_reports_dir / "RunInfo.xml"
 
 
-# @pytest.fixture(name="bcl2fastq_demux_stats_files")
-# def fixture_bcl2fastq_demux_stats_files(
-#     bcl2fastq_conversion_stats: Path, bcl2fastq_demultiplexing_stats: Path
-# ) -> Dict:
-#     """return a fixture for bcl2fastq demux_stats_files"""
-#     return {
-#         "conversion_stats": bcl2fastq_conversion_stats,
-#         "demultiplexing_stats": bcl2fastq_demultiplexing_stats,
-#         "runinfo": "",
-#     }
-
-
 @pytest.fixture(name="dragen_demux_stats_files")
 def fixture_dragen_demux_stats_files(
     dragen_adapter_metrics: Path, dragen_demultiplex_stats: Path, dragen_run_info: Path
@@ -142,25 +113,6 @@ def fixture_dragen_demux_stats_files(
         "demultiplexing_stats": dragen_demultiplex_stats,
         "runinfo": dragen_run_info,
     }
-
-
-@pytest.fixture(name="bcl2fastq_flowcell_object")
-def fixture_bcl2fastq_flowcell_object(bcl2fastq_flowcell_dir: Path) -> Flowcell:
-    return Flowcell(bcl2fastq_flowcell_dir)
-
-
-@pytest.fixture(name="bcl2fastq_demux_results")
-def fixture_bcl2fastq_demux_results(
-    demultiplexed_flowcell: Path,
-    bcl2fastq_flowcell_object: Flowcell,
-    bcl2fastq_demux_stats_files: Dict,
-) -> DemuxResults:
-    return DemuxResults(
-        demux_dir=demultiplexed_flowcell,
-        flowcell=bcl2fastq_flowcell_object,
-        bcl_converter="bcl2fastq",
-        demux_stats_files=bcl2fastq_demux_stats_files,
-    )
 
 
 @pytest.fixture(name="dragen_demux_results")
